// src/domain/types.ts
export type MidiTargetParameter = "ToggleBypass" | "WahPedalPosition" | "DelayTime" | "DelayLevel" | "DistortionLevel" | "DistortionThreshold";